Move EphemeralKey
