+++ example/src/main/java/com/stripe/example/paymentsheet/PaymentSheetViewModel.kt
@@ -10,7 +10,6 @@ import com.stripe.example.module.BackendApiFactory
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.single
 import kotlinx.coroutines.flow.singleOrNull
-import org.json.JSONObject
 import kotlin.coroutines.CoroutineContext
 
 internal class PaymentSheetViewModel(
@@ -49,25 +48,6 @@ internal class PaymentSheetViewModel(
 
             inProgress.postValue(false)
             emit(remoteEphemeralKeyResult.getOrNull())
-        }
-    }
-
-    internal data class EphemeralKey(
-        val key: String,
-        val customer: String
-    ) {
-        companion object {
-            fun fromJson(json: JSONObject): EphemeralKey {
-                val secret = json.getString("secret")
-                val associatedObjectArray = json.getJSONArray("associated_objects")
-                val typeObject = associatedObjectArray.getJSONObject(0)
-                val customerId = typeObject.getString("id")
-
-                return EphemeralKey(
-                    key = secret,
-                    customer = customerId
-                )
-            }
         }
     }
