+++ example/src/main/java/com/stripe/example/paymentsheet/Repository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 internal interface Repository {
     fun clearKeys()
 
-    suspend fun fetchLocalEphemeralKey(): Flow<PaymentSheetViewModel.EphemeralKey?>
+    suspend fun fetchLocalEphemeralKey(): Flow<EphemeralKey?>
 
-    suspend fun fetchRemoteEphemeralKey(): Flow<Result<PaymentSheetViewModel.EphemeralKey>>
+    suspend fun fetchRemoteEphemeralKey(): Flow<Result<EphemeralKey>>
 }
